% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\docType{data}
\name{density_plot}
\alias{density_plot}
\title{A plot
Create density plot in ggplot2.}
\format{
List of 9 (ggplot format)
}
\usage{
data(density_plot)
}
\description{
A plot
Create density plot in ggplot2.
}
\examples{
data(density_plot)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
